Voice and chat works now as it should.
The ingame voice works compleatly now as intendet. Players as ghost can
use the voice as if they would just be dead.

The chat on the other hand is a little tricky and needs the simple chat
processor, in the mean time with this commit/version ghosts can talk only
to dead players on the same team. Also they can only read the chat of the
living.